% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{wf_home}
\alias{wf_home}
\title{#' build_workflower_params
build_workflower_params <- function() {
  #' update_package
  #' importFrom fs file_exists
  params <- list(
    styler = c(
      directories = 'R',
      files = 'main.R'
    )
  )
}
wf_home}
\usage{
wf_home()
}
\description{
#' build_workflower_params
build_workflower_params <- function() {
  #' update_package
  #' importFrom fs file_exists
  params <- list(
    styler = c(
      directories = 'R',
      files = 'main.R'
    )
  )
}
wf_home
}
